+++ petlife-services/petlife-services-boot/src/main/java/br/com/duti/petlife/repository/interfaces/IAppointmentTypeRepository.java
@@ -7,6 +7,6 @@ import br.com.duti.petlife.models.AppointmentType;
 public interface IAppointmentTypeRepository {
 
 	List<AppointmentType> getAll();
-	AppointmentType findById(Class<?> obj, final Long id);
+	AppointmentType findById(Class<?> obj, final Integer id);
 
 }
